Bring in gpcc_scan_filenode as gp_scan_filenode
The GPCC team put together an excellent function for displaying the true
on-disk size of a table.  This is especially helpful for AO tables
because the current pg_table_size uses the position of EOF, which can be
inaccurate in the presence of aborted transactions.

This commit reimplements the same function, as-is, in core gpdb code.