Update behavior of size calculation for AO tables
Historically, AO tables have reported only logical size, which
under-reports compared to physical size in some scenarios (such as in
the presence of aborted transactions).  This also contradicts how heap
table size is reported. This commit adds custom logic for statting all
files for a given AO relation, to better align with expected behavior.

This commit adds an additional behavior change. For AO tables, we do
not have the same file storage layout, and in particular we do not have
_fsm or _vm forks. The pg_relation_size function allows the
specification of a target fork to report the size of. This commit adds
another signature for pg_relation_size, which will accept a boolean
second argument. This boolean will indicate whether to include the AO
auxiliary tables in the reported size, as these are (roughly) equivalent
to the non-main forks of heap tables. This will be defaulted to false in
the case with only one argument provided, to align with the behavior of
reporting only the main fork for a heap table.

Co-authored-by: Divyesh Vanjare <vanjared@vmware.com>